Fixed abstract classes' names, thanks codacy
