+++ demo/src/main/java/com/eis/geoCalendar/demo/MapsActivity.java
@@ -7,8 +7,8 @@ import androidx.fragment.app.FragmentActivity;
 
 import com.eis.geoCalendar.app.GenericEventManager;
 import com.eis.geoCalendar.demo.Behaviour.EventMapBehaviour;
-import com.eis.geoCalendar.demo.Dialogs.AddMapEventDialog;
-import com.eis.geoCalendar.demo.Dialogs.RemoveMapEventDialog;
+import com.eis.geoCalendar.demo.Dialogs.AddEventDialog;
+import com.eis.geoCalendar.demo.Dialogs.RemoveEventDialog;
 import com.eis.geoCalendar.demo.Localization.LocationManager;
 import com.eis.geoCalendar.events.Event;
 import com.eis.geoCalendar.events.EventManager;
@@ -44,8 +44,8 @@ public class MapsActivity extends FragmentActivity {
 
         //Creating the "Behaviour Manager"
         eventMapBehaviour = new EventMapBehaviour<>(getApplicationContext());
-        eventMapBehaviour.setAddEventDialog(new AddMapEventDialog());
-        eventMapBehaviour.setRemoveEventDialog(new RemoveMapEventDialog());
+        eventMapBehaviour.setAddEventDialog(new AddEventDialog());
+        eventMapBehaviour.setRemoveEventDialog(new RemoveEventDialog());
         eventMapBehaviour.setLocationRetriever(locationManager);
         eventMapBehaviour.setSupportFragmentManager(getSupportFragmentManager());
